Add files via upload
A user can register and log in to the system. If the user forgets their password, they can recover access by setting a new password. After logging in, the user has the option to update their profile. Additionally, the user can search for jobs based on their skills and apply for the desired positions.